Feat seperated mock data from auth test
