+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "ES2017",
     "module": "Node16",
-    "rootDir": "src",
+    "rootDir": ".",
     "outDir": "dist",
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
@@ -16,9 +16,9 @@
     "strictPropertyInitialization": true,
     "experimentalDecorators": true
   },
+  "include": ["src/**/*", "test/**/*"],
+  "exclude": ["node_modules", "jest.config.js"],
   "ts-node": {
     "files": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules", "jest.config.js"]
+  }
 }
